#374 - Expose the Android API used by the Flutter plugin as a standalone class for usage in Android native code
