+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/FlutterSecureStoragePlugin.java
@@ -1,30 +1,15 @@
 package com.it_nomads.fluttersecurestorage;
 
 import android.content.Context;
-import android.content.SharedPreferences;
-import android.os.Build;
 import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.Looper;
-import android.security.keystore.KeyGenParameterSpec;
-import android.security.keystore.KeyProperties;
-import android.util.Base64;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.RequiresApi;
-import androidx.security.crypto.EncryptedSharedPreferences;
-import androidx.security.crypto.MasterKey;
 
-import com.it_nomads.fluttersecurestorage.ciphers.StorageCipher;
-import com.it_nomads.fluttersecurestorage.ciphers.StorageCipher18Implementation;
-
-import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.nio.charset.Charset;
-import java.security.GeneralSecurityException;
-import java.util.HashMap;
 import java.util.Map;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
@@ -37,25 +22,14 @@ import io.flutter.plugin.common.MethodChannel.Result;
 public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlugin {
 
     private static final String TAG = "FlutterSecureStoragePl";
-    private static final String ELEMENT_PREFERENCES_KEY_PREFIX = "VGhpcyBpcyB0aGUgcHJlZml4IGZvciBhIHNlY3VyZSBzdG9yYWdlCg";
-    private static final String SHARED_PREFERENCES_NAME = "FlutterSecureStorage";
     private MethodChannel channel;
-    private SharedPreferences preferences;
-    private SharedPreferences nonEncryptedPreferences;
-    private Charset charset;
-    private StorageCipher storageCipher;
-    // Necessary for deferred initialization of storageCipher.
-    private Context applicationContext;
+    private SecureStorageAndroid secureStorage;
     private HandlerThread workerThread;
     private Handler workerThreadHandler;
-    private boolean useEncryptedSharedPreferences = false;
-    private boolean resetOnError = false;
 
     public void initInstance(BinaryMessenger messenger, Context context) {
         try {
-            applicationContext = context.getApplicationContext();
-            nonEncryptedPreferences = context.getSharedPreferences(SHARED_PREFERENCES_NAME, Context.MODE_PRIVATE);
-            charset = Charset.forName("UTF-8");
+            secureStorage = new SecureStorageAndroid(context);
 
             workerThread = new HandlerThread("com.it_nomads.fluttersecurestorage.worker");
             workerThread.start();
@@ -66,75 +40,6 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
         } catch (Exception e) {
             Log.e(TAG, "Registration failed", e);
         }
-    }
-
-    private void checkAndMigrateToEncrypted(SharedPreferences source, SharedPreferences target) {
-        for (Map.Entry<String, ?> entry : source.getAll().entrySet()) {
-            Object v = entry.getValue();
-            String key = entry.getKey();
-            if (v instanceof String && key.contains(ELEMENT_PREFERENCES_KEY_PREFIX))
-                try {
-                    final String decodedValue = decodeRawValue((String) v);
-                    target.edit().putString(key, (decodedValue)).apply();
-                    source.edit().remove(key).apply();
-                } catch (Exception e) {
-                    Log.e(TAG, "Data migration failed", e);
-                }
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private void ensureInitialized(Map<String, Object> arguments) {
-        Map<String, Object> options = (Map<String, Object>) arguments.get("options");
-        if (options != null) {
-            useEncryptedSharedPreferences = useEncryptedSharedPreferences(options);
-            resetOnError = resetOnError(options);
-
-            if (storageCipher == null) {
-                try {
-                    storageCipher = new StorageCipher18Implementation(applicationContext);
-                } catch (Exception e) {
-                    Log.e(TAG, "StorageCipher initialization failed", e);
-                }
-            }
-
-            if (useEncryptedSharedPreferences &&
-                    Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-
-                try {
-                    preferences = initializeEncryptedSharedPreferencesManager(applicationContext);
-                } catch (Exception e) {
-                    Log.e("FlutterSecureStoragePl", "EncryptedSharedPreferences initialization failed", e);
-                }
-
-                checkAndMigrateToEncrypted(nonEncryptedPreferences, preferences);
-            } else {
-                preferences = nonEncryptedPreferences;
-            }
-        }
-    }
-
-    private boolean resetOnError(Map<String, Object> arguments) {
-        return arguments.containsKey("resetOnError") && arguments.get("resetOnError").equals("true");
-    }
-
-    private boolean useEncryptedSharedPreferences(Map<String, Object> arguments) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) return false;
-        return arguments.containsKey("encryptedSharedPreferences") && arguments.get("encryptedSharedPreferences").equals("true");
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.M)
-    private SharedPreferences initializeEncryptedSharedPreferencesManager(Context context) throws GeneralSecurityException, IOException {
-        MasterKey key = new MasterKey.Builder(context)
-                .setKeyGenParameterSpec(
-                        new KeyGenParameterSpec
-                                .Builder(MasterKey.DEFAULT_MASTER_KEY_ALIAS, KeyProperties.PURPOSE_ENCRYPT | KeyProperties.PURPOSE_DECRYPT)
-                                .setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_NONE)
-                                .setBlockModes(KeyProperties.BLOCK_MODE_GCM)
-                                .setKeySize(256).build())
-                .build();
-        return EncryptedSharedPreferences.create(context, SHARED_PREFERENCES_NAME, key, EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
-                EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM);
     }
 
     @Override
@@ -151,6 +56,7 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
             channel.setMethodCallHandler(null);
             channel = null;
         }
+        secureStorage = null;
     }
 
     @Override
@@ -160,77 +66,28 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
         workerThreadHandler.post(new MethodRunner(call, result));
     }
 
+    @SuppressWarnings({"unchecked", "ConstantConditions"})
+    private boolean getResetOnErrorFromCall(MethodCall call) {
+        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
+        return arguments.containsKey("resetOnError") && arguments.get("resetOnError").equals("true");
+    }
+
+    @SuppressWarnings({"unchecked", "ConstantConditions"})
+    private boolean getUseEncryptedSharedPreferencesFromCall(MethodCall call) {
+        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
+        return arguments.containsKey("encryptedSharedPreferences") && arguments.get("encryptedSharedPreferences").equals("true");
+    }
+
     @SuppressWarnings("unchecked")
     private String getKeyFromCall(MethodCall call) {
         Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-        String rawKey = (String) arguments.get("key");
-        return addPrefixToKey(rawKey);
+        return (String) arguments.get("key");
     }
 
     @SuppressWarnings("unchecked")
-    private Map<String, String> readAll(boolean useEncryptedSharedPreference) throws Exception {
-        Map<String, String> raw = (Map<String, String>) preferences.getAll();
-
-        Map<String, String> all = new HashMap<>();
-        for (Map.Entry<String, String> entry : raw.entrySet()) {
-            String keyWithPrefix = entry.getKey();
-            if (keyWithPrefix.contains(ELEMENT_PREFERENCES_KEY_PREFIX)) {
-                String key = entry.getKey().replaceFirst(ELEMENT_PREFERENCES_KEY_PREFIX + '_', "");
-                if (useEncryptedSharedPreference) {
-                    all.put(key, entry.getValue());
-                } else {
-                    String rawValue = entry.getValue();
-                    String value = decodeRawValue(rawValue);
-
-                    all.put(key, value);
-                }
-            }
-        }
-        return all;
-    }
-
-    private void deleteAll() {
-        preferences.edit().clear().apply();
-    }
-
-    private void write(String key, String value, boolean useEncryptedSharedPreference) throws Exception {
-        SharedPreferences.Editor editor = preferences.edit();
-
-        if (useEncryptedSharedPreference) {
-            editor.putString(key, value);
-        } else {
-            byte[] result = storageCipher.encrypt(value.getBytes(charset));
-            editor.putString(key, Base64.encodeToString(result, 0));
-        }
-        editor.apply();
-    }
-
-    private String read(String key, boolean useEncryptedSharedPreference) throws Exception {
-        String rawValue = preferences.getString(key, null);
-        if (useEncryptedSharedPreference) {
-            return rawValue;
-        }
-        return decodeRawValue(rawValue);
-    }
-
-    private void delete(String key) {
-        SharedPreferences.Editor editor = preferences.edit();
-        editor.remove(key);
-        editor.apply();
-    }
-
-    private String addPrefixToKey(String key) {
-        return ELEMENT_PREFERENCES_KEY_PREFIX + "_" + key;
-    }
-
-    private String decodeRawValue(String value) throws Exception {
-        if (value == null) {
-            return null;
-        }
-        byte[] data = Base64.decode(value, 0);
-        byte[] result = storageCipher.decrypt(data);
-
-        return new String(result, charset);
+    private String getValueFromCall(MethodCall call) {
+        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
+        return (String) arguments.get("value");
     }
 
     /**
@@ -274,19 +131,19 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
         }
 
         @Override
-        @SuppressWarnings("unchecked")
         public void run() {
+            boolean resetOnError = false;
+            boolean useEncryptedSharedPreferences = false;
             try {
+                resetOnError = getResetOnErrorFromCall(call);
+                useEncryptedSharedPreferences = getUseEncryptedSharedPreferencesFromCall(call);
                 switch (call.method) {
                     case "write": {
                         String key = getKeyFromCall(call);
-                        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-                        ensureInitialized(arguments);
-
-                        String value = (String) arguments.get("value");
+                        String value = getValueFromCall(call);
 
                         if (value != null) {
-                            write(key, value, useEncryptedSharedPreferences);
+                            secureStorage.write(key, value, useEncryptedSharedPreferences);
                             result.success(null);
                         } else {
                             result.error("null", null, null);
@@ -295,11 +152,9 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
                     }
                     case "read": {
                         String key = getKeyFromCall(call);
-                        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-                        ensureInitialized(arguments);
 
-                        if (preferences.contains(key)) {
-                            String value = read(key, useEncryptedSharedPreferences);
+                        if (secureStorage.containsKey(key, useEncryptedSharedPreferences)) {
+                            String value = secureStorage.read(key, useEncryptedSharedPreferences);
                             result.success(value);
                         } else {
                             result.success(null);
@@ -307,36 +162,26 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
                         break;
                     }
                     case "readAll": {
-                        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-                        ensureInitialized(arguments);
-
-                        Map<String, String> value = readAll(useEncryptedSharedPreferences);
+                        Map<String, String> value = secureStorage.readAll(useEncryptedSharedPreferences);
                         result.success(value);
                         break;
                     }
                     case "containsKey": {
                         String key = getKeyFromCall(call);
-                        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-                        ensureInitialized(arguments);
 
-                        boolean containsKey = preferences.contains(key);
+                        boolean containsKey = secureStorage.containsKey(key, useEncryptedSharedPreferences);
                         result.success(containsKey);
                         break;
                     }
                     case "delete": {
                         String key = getKeyFromCall(call);
-                        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-                        ensureInitialized(arguments);
 
-                        delete(key);
+                        secureStorage.delete(key, useEncryptedSharedPreferences);
                         result.success(null);
                         break;
                     }
                     case "deleteAll": {
-                        Map<String, Object> arguments = (Map<String, Object>) call.arguments;
-                        ensureInitialized(arguments);
-
-                        deleteAll();
+                        secureStorage.deleteAll(useEncryptedSharedPreferences);
                         result.success(null);
                         break;
                     }
@@ -347,7 +192,7 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
 
             } catch (Exception e) {
                 if (resetOnError) {
-                    deleteAll();
+                    secureStorage.deleteAll(useEncryptedSharedPreferences);
                     result.success("Data has been reset");
                 } else {
                     StringWriter stringWriter = new StringWriter();
